Atualização dia 30/03/2025 proxima atualização verificar porque4 não está salvando o prontuario no banco de dados, está retornando que está com ID duplicado, tratar o erro, deve contem 1 prontuario para um paciente e o prontuario pode conter varios ProfissionalDeSaude e varias consultas e varias Prescricao

+++ src/main/java/com/SSHSS/Repository/ExameRepository.java
@@ -3,5 +3,5 @@ package com.SSHSS.Repository;
 import com.SSHSS.Model.Exame;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExameRepository extends JpaRepository<Exame, Integer> {
+public interface ExameRepository extends JpaRepository<Exame, Long> {
 }
