+++ src/main/java/com/SSHSS/Repository/PrescricaoRepository.java
@@ -3,5 +3,5 @@ package com.SSHSS.Repository;
 import com.SSHSS.Model.Prescricao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrescricaoRepository extends JpaRepository<Prescricao, Integer> {
+public interface PrescricaoRepository extends JpaRepository<Prescricao, Long> {
 }
